+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.2.12
+current_version = 0.0.1
 commit = True
 tag = True
 
@@ -7,7 +7,7 @@ tag = True
 search = version='{current_version}'
 replace = version='{new_version}'
 
-[bumpversion:file:ocean_py/__init__.py]
+[bumpversion:file:starfish_py/__init__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
