+++ docs/source/conf.py
@@ -21,9 +21,9 @@ sys.path.insert(0, os.path.abspath('../../'))
 
 # -- Project information -----------------------------------------------------
 
-project = 'ocean-py'
-copyright = 'ocean-py contributors'
-author = 'ocean-py contributors'
+project = 'starfish-py'
+copyright = 'starfish-py contributors'
+author = 'starfish-py contributors'
 
 # The full version, including alpha/beta/rc tags
 release = '0.0.1'
@@ -44,7 +44,7 @@ extensions = [
 
 # apidoc settings
 # See https://github.com/sphinx-contrib/apidoc
-apidoc_module_dir = '../../ocean_py'
+apidoc_module_dir = '../../starfish_py'
 # apidoc_output_dir = 'api' by default, and leave it that way!
 apidoc_separate_modules = True
 # See https://www.sphinx-doc.org/en/master/man/sphinx-apidoc.html
