+++ docs/source/index.rst
@@ -1,4 +1,4 @@
-ocean-py API Reference Docs
+starfish-py API Reference Docs
 ===========================
 
 These docs are API Reference docs.
