+++ starfish_py/models/squid_model.py
@@ -2,7 +2,7 @@
     SquidModel - Access squid services using the squid-py api
 """
 
-from ocean_py.models.model_base import ModelBase
+from starfish_py.models.model_base import ModelBase
 from squid_py.service_agreement.utils import (
     get_sla_template_path,
     register_service_agreement_template
@@ -13,7 +13,7 @@ from squid_py import (
     ServiceTypes
 )
 
-# from ocean_py import logger
+# from starfish_py import logger
 
 class SquidModel(ModelBase):
     def __init__(self, ocean):
