+++ tests/resources/config.ini
@@ -1,4 +1,4 @@
-[ocean-py]
+[starfish-py]
 ocean_url = http://localhost:8545
 contract_path = artifacts
