+++ tox.ini
@@ -8,7 +8,7 @@ python =
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 ocean-py
+commands = flake8 starfish-py
 
 [testenv]
 # passenv = CODACY_PROJECT_TOKEN
@@ -24,7 +24,7 @@ commands =
     {toxinidir}/scripts/wait_for_migration_and_extract_keeper_artifacts.sh
     pip install -U pip
     pip list
-    coverage run --source ocean_py -m py.test  --basetemp={envtmpdir}
+    coverage run --source starfish_py -m py.test  --basetemp={envtmpdir}
     coverage report
     coverage xml
 #    python-codacy-coverage -r coverage.xml
